requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
click==7.1.2
Django==3.0.8
django-crispy-forms==1.9.2
form==0.0.1
pathspec==0.8.0
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
